+++ internal/scheduledupgrades/error.go
@@ -1,0 +1,7 @@
+package scheduledupgrades
+
+import "github.com/giantswarm/microerror"
+
+var notAllowedError = &microerror.Error{
+	Kind: "notAllowedError",
+}